css/css-grid-1/alignment/row-axis/grid-row-axis-alignment-positioned-items-vertical-rl-004.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Grid Layout Test: Self-Alignment along row axis of absolute positioned items with well defined grid positions in a vertical-rl grid</title>
<link rel="author" title="Javier Fernandez Garcia-Boente" href="mailto:jfernandez@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#row-align">
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#abspos-items">
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-justify-self">
<link rel="help" href="https://drafts.csswg.org/css-align/#valdef-self-position-start">
<meta name="assert" content="Absolute positioned grid items with 'right' value for justify-self are flushed with its alignment container's 'right' edge.">
<style>
.grid {
  position: relative;
  display: inline-grid;
  grid-template-columns: 100px 150px;
  grid-template-rows: 150px 100px;
  font: 10px/1 Ahem;
  background: grey;
  align-items: start;
  writing-mode: vertical-rl;
}
.grid > div { position: absolute; }
.grid > :nth-child(1) { background: green; }
.grid > :nth-child(2) { background: blue; }
.grid > :nth-child(3) { background: yellow; }
.grid > :nth-child(4) { background: red; }

.firstRowFirstColumn {
  grid-row: 1 / 2;
  grid-column: 1 / 2;
  justify-self: right;
}
.firstRowSecondColumn {
  grid-row: 1 / 2;
  grid-column: 2 / 3;
  justify-self: right;
}
.secondRowFirstColumn {
  grid-row: 2 / 3;
  grid-column: 1 / 2;
  justify-self: right;
}
.secondRowSecondColumn {
  grid-row: 2 / 3;
  grid-column: 2 / 3;
  justify-self: right;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../support/check-layout-th.js"></script>
<body onload="checkLayout('.grid')">
<div class="grid">
  <div data-offset-x="240" data-offset-y="40"  data-expected-width="10" data-expected-height="60" class="firstRowFirstColumn">X XX X</div>
  <div data-offset-x="220" data-offset-y="180" data-expected-width="30" data-expected-height="70" class="firstRowSecondColumn">XX X<br>X XXX X<br>XX XXX</div>
  <div data-offset-x="90"  data-offset-y="40"  data-expected-width="10" data-expected-height="60" class="secondRowFirstColumn">X XX X</div>
  <div data-offset-x="60"  data-offset-y="190" data-expected-width="40" data-expected-height="60" class="secondRowSecondColumn">XX X<br>X XXX<br>X<br>XX XXX</div>
</div>
</body>
